# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MediTrack Senior API
  description: |
    API REST del backend de MediTrack Senior.

    Autenticacion mediante JWT (Bearer). El token se obtiene en
    `POST /api/auth/login` y se envia en el encabezado
    `Authorization: Bearer <token>`.

    Los errores usan un unico formato (ver el esquema `ApiError`).
  version: v1
servers:
  - url: http://localhost:8080
    description: Entorno local de desarrollo
...
paths:
...
  /api/pacientes/{pacienteId}/cuidadores:
    get:
      tags: [Cuidadores]
      summary: Listar cuidadores con asignacion vigente de un paciente (solo ADMINISTRADOR)
      parameters:
        - $ref: '#/components/parameters/PacienteId'
      responses:
        '200':
          description: Lista de asignaciones vigentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AsignacionCuidador'
        '403':
          description: El usuario autenticado no es ADMINISTRADOR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags: [Cuidadores]
      summary: Asignar un cuidador a un paciente (solo ADMINISTRADOR)
      parameters:
        - $ref: '#/components/parameters/PacienteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuidadorRequest'
      responses:
        '201':
          description: Asignacion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsignacionCuidador'
        '403':
          description: El usuario autenticado no es ADMINISTRADOR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Paciente o usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: El cuidador ya tiene una asignacion vigente o no tiene el rol
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/pacientes/{pacienteId}/cuidadores/{usuarioId}:
    delete:
      tags: [Cuidadores]
      summary: Desasignar un cuidador (solo ADMINISTRADOR)
      parameters:
        - $ref: '#/components/parameters/PacienteId'
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '204':
          description: Asignacion cerrada (hasta = hoy)
        '403':
          description: El usuario autenticado no es ADMINISTRADOR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Asignacion vigente no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UsuarioId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    PacienteId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    AsignacionCuidador:
      type: object
      properties:
        id:
          type: integer
          format: int64
        usuarioId:
          type: integer
          format: int64
        pacienteId:
          type: integer
          format: int64
        desde:
          type: string
          format: date
        hasta:
          type: string
          format: date
          nullable: true
    CuidadorRequest:
      type: object
      required: [usuarioId]
      properties:
        usuarioId:
          type: integer
          format: int64
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        codigo:
          type: string
          example: VALIDACION
        mensaje:
          type: string
        detalles:
          type: array
          nullable: true
          items:
            type: string
security:
  - bearerAuth: []